fixed warning

+++ tests/cunit/test_async_mpi.c
@@ -141,8 +141,8 @@ int msg_handler(int verbose, int my_rank, int io_rank, int component_count,
                                     union_comm[cmp], &req[cmp])))
                 MPIERR(mpierr);
             if (verbose)
-                printf("my_rank %d MPI_Irecv req[%d] = %d\n", my_rank, cmp,
-                       (int)(req[cmp]));
+                printf("my_rank %d MPI_Irecv req[%d] = %lld\n", my_rank, cmp,
+                       (long long int)(req[cmp]));
         }
     }
 
@@ -159,11 +159,12 @@ int msg_handler(int verbose, int my_rank, int io_rank, int component_count,
         {
             if (verbose)
             {
-                printf("my_rank %d about to call MPI_Waitany req[0] = %d "
-                       "MPI_REQUEST_NULL = %d\n", my_rank, (int)(req[0]),
+                printf("my_rank %d about to call MPI_Waitany req[0] = %lld "
+                       "MPI_REQUEST_NULL = %d\n", my_rank, (long long int)(req[0]),
                        MPI_REQUEST_NULL);
                 for (int c = 0; c < component_count; c++)
-                    printf("my_rank %d req[%d] = %d\n", my_rank, c, (int)(req[c]));
+                    printf("my_rank %d req[%d] = %lld\n", my_rank, c,
+                           (long long int)(req[c]));
             }
             if ((mpierr = MPI_Waitany(component_count, req, &index, &status)))
                 MPIERR(mpierr);
